create a directory for task board and its components
