+++ src/components/Kanban/index.ts
@@ -2,4 +2,4 @@ export * from './CreateTickets/CreateTicket'
 export * from './Components/SelectStatus'
 export * from './TicketDetail/TicketDetail'
 export * from './TicketDetail/TicketDetailFunctions'
-export * from './TaskBoard'
+export * from './TaskBoard/TaskBoard'
