+++ src/utils/helpers/stringHelpers.ts
@@ -1,0 +1,5 @@
+export const concatenatedString = (statusString: string) =>
+  statusString.replace(/\s+/g, '')
+
+export const splitCamelCaseToWords = (statusString: string) =>
+  statusString?.split(/(?=[A-Z])/).join(' ')
